Made get_time_bounds and get_next_month public
