+++ esmvalcore/cmor/_fixes/cmip6/iitm_esm.py
@@ -3,7 +3,7 @@ import logging
 
 import numpy as np
 
-from esmvalcore.cmor._fixes.fix import get_time_bounds
+from esmvalcore.cmor.fixes import get_time_bounds
 
 from ..common import OceanFixGrid
 from ..fix import Fix
