+++ esmvalcore/cmor/_fixes/shared.py
@@ -1,6 +1,7 @@
 """Shared functions for fixes."""
 import logging
 import os
+from datetime import datetime
 from functools import lru_cache
 
 import dask.array as da
@@ -9,7 +10,10 @@ import numpy as np
 import pandas as pd
 from cf_units import Unit
 from iris import NameConstraint
+from iris.coords import Coord
 from scipy.interpolate import interp1d
+
+from esmvalcore.iris_helpers import date2num
 
 logger = logging.getLogger(__name__)
 
@@ -411,3 +415,92 @@ def fix_ocean_depth_coord(cube):
     depth_coord.units = 'm'
     depth_coord.long_name = 'ocean depth coordinate'
     depth_coord.attributes = {'positive': 'down'}
+
+
+def get_next_month(month: int, year: int) -> tuple[int, int]:
+    """Get next month and year.
+
+    Parameters
+    ----------
+    month:
+        Current month.
+    year:
+        Current year.
+
+    Returns
+    -------
+    tuple[int, int]
+        Next month and next year.
+
+    """
+    if month != 12:
+        return month + 1, year
+    return 1, year + 1
+
+
+def get_time_bounds(time: Coord, freq: str):
+    """Get bounds for time coordinate.
+
+    For monthly data, use the first day of the current month and the first day
+    of the next month. For yearly or decadal data, use 1 January of the current
+    year and 1 January of the next year or 10 years from the current year. For
+    other frequencies (daily, 6-hourly, 3-hourly, hourly), half of the
+    frequency is subtracted/added from the current point in time to get the
+    bounds.
+
+    Parameters
+    ----------
+    time:
+        Time coordinate.
+    freq:
+        Frequency.
+
+    Returns
+    -------
+    np.ndarray
+        Time bounds
+
+    Raises
+    ------
+    NotImplementedError
+        Non-supported frequency is given.
+
+    """
+    bounds = []
+    dates = time.units.num2date(time.points)
+    for step, date in enumerate(dates):
+        month = date.month
+        year = date.year
+        if freq in ['mon', 'mo']:
+            next_month, next_year = get_next_month(month, year)
+            min_bound = date2num(datetime(year, month, 1, 0, 0),
+                                 time.units, time.dtype)
+            max_bound = date2num(datetime(next_year, next_month, 1, 0, 0),
+                                 time.units, time.dtype)
+        elif freq == 'yr':
+            min_bound = date2num(datetime(year, 1, 1, 0, 0),
+                                 time.units, time.dtype)
+            max_bound = date2num(datetime(year + 1, 1, 1, 0, 0),
+                                 time.units, time.dtype)
+        elif freq == 'dec':
+            min_bound = date2num(datetime(year, 1, 1, 0, 0),
+                                 time.units, time.dtype)
+            max_bound = date2num(datetime(year + 10, 1, 1, 0, 0),
+                                 time.units, time.dtype)
+        else:
+            delta = {
+                'day': 12.0 / 24,
+                '6hr': 3.0 / 24,
+                '3hr': 1.5 / 24,
+                '1hr': 0.5 / 24,
+            }
+            if freq not in delta:
+                raise NotImplementedError(
+                    f"Cannot guess time bounds for frequency '{freq}'"
+                )
+            point = time.points[step]
+            min_bound = point - delta[freq]
+            max_bound = point + delta[freq]
+        bounds.append([min_bound, max_bound])
+
+    return np.array(bounds)
