+++ esmvalcore/cmor/fixes.py
@@ -1,8 +1,15 @@
 """Functions for fixing specific issues with datasets."""
 
-from ._fixes.shared import add_altitude_from_plev, add_plev_from_altitude
+from ._fixes.shared import (
+    add_altitude_from_plev,
+    add_plev_from_altitude,
+    get_next_month,
+    get_time_bounds,
+)
 
 __all__ = [
     'add_altitude_from_plev',
     'add_plev_from_altitude',
+    'get_time_bounds',
+    'get_next_month',
 ]
