+++ tests/unit/cmor/test_fixes.py
@@ -8,6 +8,8 @@ import esmvalcore.cmor.fixes as fixes
 @pytest.mark.parametrize('func', [
     'add_altitude_from_plev',
     'add_plev_from_altitude',
+    'get_next_month',
+    'get_time_bounds',
 ])
 def test_imports(func):
     assert func in fixes.__all__
